requirements.txt
azure.cognitiveservices.vision.computervision
msrest.authentication
array
os
PIL